fix: Add CORS preflight support and security fixes for MCP OAuth
## CORS Fix
Allow OPTIONS requests without authentication in McpAuthFilter to support
CORS preflight checks from web-based MCP clients.

This enables proper CORS flow:
1. Browser sends OPTIONS preflight
2. Server responds with CORS headers (200 OK)
3. Browser sends actual POST request with Authorization header
4. Server authenticates and processes request

Without this fix, OPTIONS requests were blocked with 401, preventing
web clients from connecting to MCP endpoints.

## Security Fixes

### Critical Security Issues Fixed:
1. **Sensitive Token Logging** (95% severity)
   - Sanitize OAuth request parameters before logging
   - Remove client_secret, code, code_verifier, refresh_token, access_token from logs
   - Prevents credential leakage in log files

2. **Token Expiry Integer Overflow** (100% severity)
   - Changed all expiry timestamps from int/Integer to long/Long
   - Fixes 2038 problem (32-bit timestamp overflow)
   - Updated: AccessToken, RefreshToken, AuthorizationCode, ConnectorOAuthProvider, OAuthTokenRepository

3. **Hardcoded Default Connector** (80% severity)
   - Made default connector configurable via MCP_DEFAULT_CONNECTOR env var
   - Defaults to null in production (requires explicit connector_name)
   - Prevents unauthorized access to test credentials in production

4. **Missing Null Checks** (85% severity)
   - Added validation for token refresh response fields
   - Validates access_token and expires_in exist before use
   - Added bounds checking for expires_in (max 1 year)

5. **Missing Input Validation** (75% severity)
   - Added connector name format validation
   - Only allows: a-z, A-Z, 0-9, _, - characters
   - Prevents path traversal and injection attacks

## Documentation
- Moved MCP docs to organized structure: openmetadata-mcp/docs/
- Created openmetadata-mcp/README.md with foundation documentation
- Moved implementation guide and testing guide to docs/ directory

## Cleanup
- Removed development test scripts (scripts/mcp-oauth-tests/)
- Removed .shell-fix-note and test-default-connector.sh
- Kept only clean final test script: test-mcp-with-token.sh

Changes:
- openmetadata-mcp/src/main/java/org/openmetadata/mcp/McpAuthFilter.java: OPTIONS CORS support
- openmetadata-mcp/src/main/java/org/openmetadata/mcp/server/transport/OAuthHttpStatelessServerTransportProvider.java: Sanitized logging
- openmetadata-mcp/src/main/java/org/openmetadata/mcp/server/auth/provider/ConnectorOAuthProvider.java: Multiple security fixes
- openmetadata-mcp/src/main/java/org/openmetadata/mcp/McpServer.java: Configurable default connector
- openmetadata-mcp/src/main/java/org/openmetadata/mcp/auth/*.java: Long timestamps
- openmetadata-mcp/src/main/java/org/openmetadata/mcp/server/auth/repository/OAuthTokenRepository.java: Long timestamps

Testing:
- OAuth flow: ✅ Working with any OAuth-enabled connector
- MCP protocol: ✅ Working via HTTP POST with JWT
- Default connector: Configurable via MCP_DEFAULT_CONNECTOR env var
- General solution: Works with ANY connector with OAuth credentials

Test command:
export MCP_DEFAULT_CONNECTOR=test-snowflake-mcp  # For testing only
./test-mcp-with-token.sh

+++ openmetadata-mcp/src/main/java/org/openmetadata/mcp/auth/AccessToken.java
@@ -13,13 +13,13 @@ public class AccessToken {
 
   private List<String> scopes;
 
-  private Integer expiresAt;
+  private Long expiresAt;
 
   private List<String> audience;
 
   public AccessToken() {}
 
-  public AccessToken(String token, String clientId, List<String> scopes, Integer expiresAt) {
+  public AccessToken(String token, String clientId, List<String> scopes, Long expiresAt) {
     this.token = token;
     this.clientId = clientId;
     this.scopes = scopes;
@@ -27,11 +27,7 @@ public class AccessToken {
   }
 
   public AccessToken(
-      String token,
-      String clientId,
-      List<String> scopes,
-      Integer expiresAt,
-      List<String> audience) {
+      String token, String clientId, List<String> scopes, Long expiresAt, List<String> audience) {
     this.token = token;
     this.clientId = clientId;
     this.scopes = scopes;
@@ -63,11 +59,11 @@ public class AccessToken {
     this.scopes = scopes;
   }
 
-  public Integer getExpiresAt() {
+  public Long getExpiresAt() {
     return expiresAt;
   }
 
-  public void setExpiresAt(Integer expiresAt) {
+  public void setExpiresAt(Long expiresAt) {
     this.expiresAt = expiresAt;
   }
